requirements.txt
Django==1.6
argparse==1.2.1
djangorestframework==2.3.14
wsgiref==0.1.2
coverage==3.7.1
psycopg2==2.5.3

